% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abfData.R
\docType{class}
\name{AbfData-class}
\alias{AbfData-class}
\title{AbfData Class}
\description{
A class storing information on current data from some ABF file,
along with some metadata for this data i.e. original file name,
sampling frequency, applied filter and number of desired states.
}
\details{
For convenience during analysis, data is saved in "long format"
(one row per measurement) and "wide format' (one row per event).
}
\section{Slots}{

\describe{
\item{\code{fileName}}{Original name of the abf file from which trace was loaded}

\item{\code{samplingFrequency}}{Sampling frequency to which the trace data was set after it was loaded}

\item{\code{rawSamplingFrequency}}{Original sampling frequency of trace data in abf-file}

\item{\code{filter}}{Filter type that was applied to data, or NA if none was applied}

\item{\code{openPoreState}}{Number denoting the automatically detected open pore state}

\item{\code{model1}}{Depmix-model used to discern blocked state from open pore states}

\item{\code{model2}}{Depmix-model used to discern different blocked states}

\item{\code{longData}}{Data frame containing trace data in a long format, i.e. one row per measured time point}

\item{\code{wideData}}{Data frame containing trace data in a wide format, i.e. one line per event}
}}

