% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abfData.R
\docType{methods}
\name{applyFilter}
\alias{applyFilter}
\alias{applyFilter,AbfData,character-method}
\title{Apply filter}
\usage{
applyFilter(abfData, filterType, cutOff, w, n, bandwidth, state)

\S4method{applyFilter}{AbfData,character}(abfData, filterType)
}
\arguments{
\item{abfData}{Object of AbfData type, containing data that is to be filtered}

\item{filterType}{Type of filter that is to be applied. Depending on the given filter type, some additional arguments may be required.
Accepted filter types: "fft" (fast fourier transform), "lowPass" (low pass Gaussian), uses signal package,"boxKernel" (moving average kernel).}

\item{cutOff}{Required for fft-filter. Determines strictness of filtering (0-1, 1= no filter, 0 = flat line)}

\item{w}{Required for lowPass-filter. Determine  reciprocal of std.dev (so larger is wider window)}

\item{n}{Required for lowPass-filter. Determines number of coefficients}

\item{bandwidth}{Required for boxKernel-filter. Determine bandwidth used for moving average (greater = smoother)}

\item{state}{Optionally, specific state at which filtering should take place}
}
\value{
AbfData-object in which  the column current.filtered, containing filtered current values is added to longData and filter slot is filled with name of used filter, but which in all other ways is identical to the given abfData object.
}
\description{
Filter white noise from the data, using the given filter type. If state is given, filter only white noise at that level.
}

