% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abfData.R
\docType{methods}
\name{detectBlockedStates}
\alias{detectBlockedStates}
\alias{detectBlockedStates,AbfData,ANY,ANY-method}
\alias{detectBlockedStates,AbfData,ANY,numeric-method}
\alias{detectBlockedStates,AbfData,numeric-method}
\alias{detectBlockedStates,AbfData-method}
\title{Detect blocked states}
\usage{
detectBlockedStates(abfData, states, splitState, manualParams, dtmin, dtmax,
  autoTrash)

\S4method{detectBlockedStates}{AbfData,ANY,numeric}(abfData, states, splitState,
  manualParams, dtmin, dtmax, autoTrash)

\S4method{detectBlockedStates}{AbfData,ANY,ANY}(abfData, states, manualParams,
  dtmin, dtmax, autoTrash)
}
\arguments{
\item{abfData}{An object of class AbfData, in which blocked and open pore state have been discerned using the detectOpenState method}

\item{states}{Integer denoting the number of blocked sub-states that should be found.}

\item{splitState}{Optionally, enter a state within which substates should be recognized.}

\item{manualParams}{A vector containing required parameters for the HMM, in the following order, for n states: prior model (n parameters), transition model (n^2), response model (2n)}

\item{dtmin}{DO NOT USE, NOT READY YET! Optional parameter, defining the minimum time a state may last in s. If adjacent states are the same, these are joined together.}

\item{dtmax}{Optional parameter, defining the maximum time a state may last in s. States lasting longer are removed}

\item{autoTrash}{set to TRUE to attempt automatically filtering out extremely high and low values (see description)}
}
\value{
An abfData-object in which either the blocked state or the given state has been split into the given number of substates.
}
\description{
A method for the determination of a given number of blocked sub-states within a blocked current state as detected using the \code{\link{detectOpenState}} method.
Optionally, provide a vector containing desired values for HMM parameters in this order:
- Initial state probabilities
- Transition probabilities, filling the transition matrix row-wise starting top left
- Intercept and standard deviation per state
If no parameters are provided, or if some parameters are set to NA, these parameters are filled in automatically using Maximum Likelihood (ML) estimations. Furthermore,
a "split state" may be defined, within which the HMM will look for the given number of substates. Any previously defined state qualifies as a potential split state.
Lastly, the autoTrash option runs a different HMM prior to the one used to discern open and closed pore states and discards the state with highest standard deviation.
This may eliminate some extreme peaks deemed artefacts, although restricting the allowed range of currents was found to work better.
}
\details{
This method is built around the depmix function from the depmixS4-package. As is stated in the description of this package, ML estimations
are made using using the nnet.default routine (nnet package) for initial state and transition model probabilities and a glm for the response model.
See the documentation of the DepmixS4-package (\url{https://cran.r-project.org/web/packages/depmixS4/depmixS4.pdf}) for further information.
}

