% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abfData.R
\docType{methods}
\name{detectOpenState}
\alias{detectOpenState}
\alias{detectOpenState,AbfData-method}
\title{Detect open pore state}
\usage{
detectOpenState(abfData, manualParams, dtmin, dtmax, autoTrash, IresMethod)

\S4method{detectOpenState}{AbfData}(abfData, manualParams, dtmin, dtmax,
  autoTrash, IresMethod = "perEvent")
}
\arguments{
\item{abfData}{an object of class AbfData}

\item{manualParams}{a vector of length ten, containing required parameters for the HMM}

\item{dtmin}{DO NOT USE, NOT READY YET! Optional parameter, defining the minimum time a state may last in s. If adjacent states are the same, these are joined together.}

\item{dtmax}{Optional parameter, defining the maximum time a state may last in s. States lasting longer are removed}

\item{autoTrash}{set to TRUE to attempt automatically filtering out extremely high and low values (see description)}

\item{IresMethod}{Select which method should be used to calculate Ires. Accepted options are "perEvent" (the default), "periodically" (per 10 mins.) and "overall"}
}
\value{
an abfData-object, with information on open and closed pore state stored in the longData and wideData dateframes. The model used for determination of those states is stored in
the 'model1' variable, as a depmix-object.
}
\description{
Method for discerning open pore state from blocked pore state, using a Hidden Markov Model (HMM). Optionally, provide a vector containing desired values for HMM parameters:
- Initial state probabilities (elements 1 and 2)
- Transition probabilities, filling the transition matrix row-wise starting top left (elements 3 to 6)
- Intercept and standard deviation for state 1 (elements 7 and 8)
- Intercept and standard deviation for state 2 (elements 9 and 10)
If no parameters are provided, or if some parameters are set to NA, these parameters are filled in automatically using Maximum Likelihood (ML) estimations.
After determining open and closed pore states, Ires is automatically caluculated and added to the longData dataframe in the abf-object. By default, Ires Is calculated by division
of current through the last open pore state. If open pore states last very short however, this may cause unrealistic fluctuation in the Ires value and one of the other options (
blocked pore state averaged per period of 10 minutes or averaged overall) may be preferred.
Lastly, the autoTrash option runs a different HMM prior to the one used to discern open and closed pore states and discards the state with highest standard deviation.
This may eliminate some extreme peaks deemed artefacts, although restricting the allowed range of currents was found to work better.
}
\details{
This method is built around the depmix function from the depmixS4-package. As is stated in the description of this package, ML estimations
are made using using the nnet.default routine (nnet package) for initial state and transition model probabilities and a glm for the response model.
See the documentation of the DepmixS4-package (\url{https://cran.r-project.org/web/packages/depmixS4/depmixS4.pdf}) for further information.
}

