% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abfData.R
\docType{methods}
\name{hist,AbfData-method}
\alias{hist,AbfData-method}
\title{Draw dwell time and current histograms}
\usage{
\S4method{hist}{AbfData}(x, bwDuration, bwCurrent, rangeDuration, rangeCurrent)
}
\arguments{
\item{x}{abfData-object of which states should be summarized in histograms}

\item{bwDuration}{Binwidth of histograms on duration. by default, binwidth is set such that entire range of values is divided in 50 bins.}

\item{bwCurrent}{Binwidth of histograms on Ires. by default, binwidth is set such that entire range of values is divided in 50 bins.}

\item{rangeDuration}{Vector of two values, denoting min and max duration displayed respectively. By default, entire range of duration values is covered.}

\item{rangeCurrent}{Vector of two values, denoting min and max Ires displayed respectively. By default, entire range of Ires values is covered.}
}
\value{
A ggplot-object containing the histogram. May be stored, modified using additional ggplot arguments or displayed immediately.
}
\description{
Overrride of histogram function, automatically produces duration and current histograms of all defined states in a given abfData object.
}
\details{
Uses the multiplot function as described in the R cookbook: \url{http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/}
}

