% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abfData.R
\name{loadHmmModels}
\alias{loadHmmModels}
\title{Load HMM Models}
\usage{
loadHmmModels(abfData, fileName)
}
\arguments{
\item{abfData}{The abfData object into which the model should be loaded}

\item{fileName}{The file name (path included if not in home directory), from which HMM models should be loaded.}
}
\value{
AbfData-object with the model1 and model2 variables set to the models in the file denoted by the given file name.
}
\description{
Load an existing HMM model to the current data in an object.
}
\note{
Running loaded existing models not supported yet!
}

