% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abfData.R
\name{newAbfData}
\alias{newAbfData}
\title{Initialize AbfData object}
\usage{
newAbfData(fileName, samplingFrequency)
}
\arguments{
\item{fileName}{Path and file name, referrring to the abf file on disk}

\item{samplingFrequency}{Optionally, sampling frequency in Hz to which trace data should be adjusted}
}
\value{
A new abfData object that stores the trace from the abf-file at the given path, adjusted sampling frequency if applicable, and metadata.
}
\description{
Initialize a new abfData object with given abf file name. Automatically loads the data into
the object.
}

