% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abfData.R
\docType{methods}
\name{restrictAxes}
\alias{restrictAxes}
\alias{restrictAxes,AbfData-method}
\title{Restrict axes}
\usage{
restrictAxes(abfData, whichSequence, minCurrent, maxCurrent, minTime, maxTime)

\S4method{restrictAxes}{AbfData}(abfData, whichSequence, minCurrent, maxCurrent,
  minTime, maxTime)
}
\arguments{
\item{abfData}{object of AbfData type}

\item{whichSequence}{Sequence on which axis restriction should be based. Accepted options are "current", "current.filtered" and "current.res"}

\item{minCurrent}{Minimum current value. points with lower current value are removed.}

\item{maxCurrent}{Maximum current value. points with higher current value are removed.}

\item{minTime}{Time in s starting from which datapoints should be retained.}

\item{maxTime}{Time in s up until which datapoints should be retained.}
}
\value{
Abf-file in which the given restrictions on time and min/max current have been applied
}
\description{
Restrict current to the range in which data points should be recognized. May solve issues
in the convergence of HMM state detection. Any combination of minimum and maximum values is allowed.
}

