% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abfData.R
\name{saveHmmModels}
\alias{saveHmmModels}
\title{Save HMM Models}
\usage{
saveHmmModels(abfData, fileName)
}
\arguments{
\item{abfData}{The abfData-object from which the models are to be saved.}

\item{fileName}{Name of the file in which the data should be stored.}
}
\description{
Save the HMM model stored in an abf object, as a file to be re-used for
similar sequences, requiring similar analysis.
}

