% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abfData.R
\docType{methods}
\name{scatterHeat}
\alias{scatterHeat}
\alias{scatterHeat,AbfData-method}
\title{Scatter-heat plot (amplitude (s.d.) vs Ires)}
\usage{
scatterHeat(abfData)

\S4method{scatterHeat}{AbfData}(abfData)
}
\arguments{
\item{abfData}{AbfData object containing the information that is to be displayed in the scatter-heat plot.}
}
\value{
A ggplot-object containing the scatter-heat plot. May be stored, modified using additional ggplot arguments or displayed immediately.
}
\description{
Method for making a scatter-heat plot (amplitude vs Ires), using the information stored in an
abfData object.
}

