% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abfData.R
\name{shinyStateFinder}
\alias{shinyStateFinder}
\title{Shiny State Finder}
\usage{
shinyStateFinder()
}
\description{
Display a Shiny-based graphical user interface, for a more intuitive application of this package.
}

